+++ YouboraLibTests/plugin/YBOptionsTest.m
@@ -40,6 +40,7 @@
     opt.networkIP = @"1.2.3.4";
     opt.networkIsp = @"isp";
     opt.networkConnectionType = @"connectiontype";
+    opt.networkObfuscateIp = @NO;
     opt.deviceCode = @"devicecode";
     opt.contentResource = @"resource";
     opt.contentIsLive = @YES;
@@ -64,6 +65,16 @@
     opt.extraparam8 = @"extra8";
     opt.extraparam9 = @"extra9";
     opt.extraparam10 = @"extra10";
+    opt.extraparam11 = @"extra11";
+    opt.extraparam12 = @"extra12";
+    opt.extraparam13 = @"extra13";
+    opt.extraparam14 = @"extra14";
+    opt.extraparam15 = @"extra15";
+    opt.extraparam16 = @"extra16";
+    opt.extraparam17 = @"extra17";
+    opt.extraparam18 = @"extra18";
+    opt.extraparam19 = @"extra19";
+    opt.extraparam20 = @"extra20";
     
     [opt encodeWithCoder:archiver];
     
@@ -107,6 +118,17 @@
     XCTAssertEqualObjects(opt.extraparam8, opt2.extraparam8);
     XCTAssertEqualObjects(opt.extraparam9, opt2.extraparam9);
     XCTAssertEqualObjects(opt.extraparam10, opt2.extraparam10);
+    XCTAssertEqualObjects(opt.extraparam11, opt2.extraparam11);
+    XCTAssertEqualObjects(opt.extraparam12, opt2.extraparam12);
+    XCTAssertEqualObjects(opt.extraparam13, opt2.extraparam13);
+    XCTAssertEqualObjects(opt.extraparam14, opt2.extraparam14);
+    XCTAssertEqualObjects(opt.extraparam15, opt2.extraparam15);
+    XCTAssertEqualObjects(opt.extraparam16, opt2.extraparam16);
+    XCTAssertEqualObjects(opt.extraparam17, opt2.extraparam17);
+    XCTAssertEqualObjects(opt.extraparam18, opt2.extraparam18);
+    XCTAssertEqualObjects(opt.extraparam19, opt2.extraparam19);
+    XCTAssertEqualObjects(opt.extraparam20, opt2.extraparam20);
+    XCTAssertEqualObjects(opt.networkObfuscateIp, opt2.networkObfuscateIp);
 }
 
 @end